+++ package.json
@@ -20,6 +20,7 @@
     "graphql": "^14.2.1",
     "graphql-import": "^0.7.1",
     "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.11",
     "mongoose": "^5.5.7"
   }
 }
